Extract monitoring to separate file
