+++ api/package.json
@@ -32,6 +32,7 @@
     "https": "^1.0.0",
     "morgan": "^1.10.0",
     "nextcloud-link": "^1.2.3",
+    "on-finished": "^2.3.0",
     "pg-promise": "^10.5.2",
     "prom-client": "^12.0.0",
     "rollup": "^2.18.0",
